feat(team-settings): consolidate Invite + BulkInvite into one card with mode toggle (#41 C3-merge)
Two separate cards on the Team Settings page (Invite New Member single
+ BulkInviteCard for paste-many) was a duplicate affordance. Day-2 task
is daily roster mgmt (Members card, just promoted in #53); the cold-open
invite UX should be one card, not two pushing the Members list further
down the page.

Implementation
- New file: team/InviteCard.tsx (~290 lines)
  - Owns single + bulk state, mode toggle
  - Single mode: email input + role select + RoleHelpPopover + Send
  - Bulk mode: textarea + CSV parsing + live error summary + batch send
  - Mode toggle is a segmented control (role="tablist") in the card header
  - Receives `cardRef` and `emailInputRef` from parent so the existing
    ?focus=invite deep-link continues to scroll-and-autofocus
- TeamSettings.tsx
  - Removes inline single-invite JSX + state (inviteEmail, inviteRole,
    isInviting) + handleInvite()
  - Removes BulkInviteCard import + render
  - Mounts a single <InviteCard> with refs and onInvitesSent={loadInvites}
  - Trims now-unused lucide imports (Send, Loader2, Shield, RotateCcw)
- Deletes team/BulkInviteCard.tsx (lifted into InviteCard's bulk mode)

Behavioural deltas
- The role select keeps its RoleHelpPopover trigger (added in #54)
- The ?focus=invite deep-link (added in #51 phase 1) still scrolls
  to the card and autofocuses the email input — refs are now passed
  as props instead of declared locally
- Bulk mode preserves all of BulkInviteCard's parsing rules: header
  row skip, # / // comment lines, [,;\t] separator, dedupe by email,
  EMAIL_REGEX validation, VALID_ROLES set
- Toast messages and email-failure copy preserved verbatim from both
  previous code paths

What this is NOT
- Not C3-files (CSV file upload + drag-drop). That's a separate
  follow-up that wires a <input type="file"> into the existing paste
  path. Out of scope for this PR
- Not a refactor of the underlying workspaceService.inviteMember
  contract — same calls, same args

Closes part of #41 (C3-merge).
